+++ tsdb/engine/tsm1/ring.go
@@ -3,7 +3,6 @@ package tsm1
 import (
 	"fmt"
 	"sync"
-	"sync/atomic"
 
 	"github.com/cespare/xxhash"
 	"github.com/influxdata/influxdb/v2/pkg/bytesutil"
@@ -20,11 +19,11 @@ const partitions = 16
 // ring is implemented as a crude hash ring, in so much that you can have
 // variable numbers of members in the ring, and the appropriate member for a
 // given series key can always consistently be found. Unlike a true hash ring
-// though, this ring is not resizeable—there must be at most 256 members in the
+// though, this ring is not resizeable—there must be at most 16 members in the
 // ring, and the number of members must always be a power of 2.
 //
 // ring works as follows: Each member of the ring contains a single store, which
-// contains a map of series keys to entries. A ring always has 256 partitions,
+// contains a map of series keys to entries. A ring always has 16 partitions,
 // and a member takes up one or more of these partitions (depending on how many
 // members are specified to be in the ring)
 //
@@ -32,12 +31,6 @@ const partitions = 16
 // key is hashed and the first 8 bits are used as an index to the ring.
 //
 type ring struct {
-	// Number of keys within the ring. This is used to provide a hint for
-	// allocating the return values in keys(). It will not be perfectly accurate
-	// since it doesn't consider adding duplicate keys, or trying to remove non-
-	// existent keys.
-	keysHint int64
-
 	// The unique set of partitions in the ring.
 	// len(partitions) <= len(continuum)
 	partitions []*partition
@@ -47,11 +40,14 @@ type ring struct {
 // power of 2, and for performance reasons should be larger than the number of
 // cores on the host. The supported set of values for n is:
 //
-//     {1, 2, 4, 8, 16, 32, 64, 128, 256}.
+//     {1, 2, 4, 8, 16}.
 //
 func newring(n int) (*ring, error) {
 	if n <= 0 || n > partitions {
-		return nil, fmt.Errorf("invalid number of paritions: %d", n)
+		return nil, fmt.Errorf("invalid number of partitions: %d", n)
+	}
+	if n&(n-1) != 0 {
+		return nil, fmt.Errorf("partitions %d is not a power of two", n)
 	}
 
 	r := ring{
@@ -78,7 +74,6 @@ func (r *ring) reset() {
 	for _, partition := range r.partitions {
 		partition.reset()
 	}
-	r.keysHint = 0
 }
 
 // getPartition retrieves the hash ring partition associated with the provided
@@ -100,25 +95,16 @@ func (r *ring) write(key []byte, values Values) (bool, error) {
 	return r.getPartition(key).write(key, values)
 }
 
-// add adds an entry to the ring.
-func (r *ring) add(key []byte, entry *entry) {
-	r.getPartition(key).add(key, entry)
-	atomic.AddInt64(&r.keysHint, 1)
-}
-
 // remove deletes the entry for the given key.
 // remove is safe for use by multiple goroutines.
 func (r *ring) remove(key []byte) {
 	r.getPartition(key).remove(key)
-	if r.keysHint > 0 {
-		atomic.AddInt64(&r.keysHint, -1)
-	}
 }
 
 // keys returns all the keys from all partitions in the hash ring. The returned
 // keys will be in order if sorted is true.
 func (r *ring) keys(sorted bool) [][]byte {
-	keys := make([][]byte, 0, atomic.LoadInt64(&r.keysHint))
+	keys := make([][]byte, 0)
 	for _, p := range r.partitions {
 		keys = append(keys, p.keys()...)
 	}
@@ -129,6 +115,8 @@ func (r *ring) keys(sorted bool) [][]byte {
 	return keys
 }
 
+// count returns the number of values in the ring
+// count is not accurate since it doesn't use read lock when iterating over partitions
 func (r *ring) count() int {
 	var n int
 	for _, p := range r.partitions {
@@ -202,7 +190,6 @@ func (r *ring) applySerial(f func([]byte, *entry) error) error {
 }
 
 func (r *ring) split(n int) []storer {
-	var keys int
 	storers := make([]storer, n)
 	for i := 0; i < n; i++ {
 		storers[i], _ = newring(len(r.partitions))
@@ -211,7 +198,6 @@ func (r *ring) split(n int) []storer {
 	for i, p := range r.partitions {
 		r := storers[i%n].(*ring)
 		r.partitions[i] = p
-		keys += len(p.store)
 	}
 	return storers
 }
@@ -259,13 +245,6 @@ func (p *partition) write(key []byte, values Values) (bool, error) {
 
 	p.store[string(key)] = e
 	return true, nil
-}
-
-// add adds a new entry for key to the partition.
-func (p *partition) add(key []byte, entry *entry) {
-	p.mu.Lock()
-	p.store[string(key)] = entry
-	p.mu.Unlock()
 }
 
 // remove deletes the entry associated with the provided key.